requirements.txt
wheel
tqdm
numpy
pathlib
scipy
scikit-image
opencv-python
mtcnn
tensorflow-gpu
keras
protobuf
